+++ src/auth/auth.module.ts
@@ -1,6 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
+import { AuthGuard } from './auth.guard';
+import { ConditionalAuthGuard } from './conditional-auth.guard';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
@@ -15,14 +17,14 @@ import { EmailModule } from '../email/email.module';
       useFactory: (configService: ConfigService) => ({
         global: true,
         secret: configService.get<string>('ACCESS_TOKEN_SECRET'),
-        signOptions: { expiresIn: '7d' },
+        signOptions: { expiresIn: '365d' },
       }),
     }),
     MikroOrmModule.forFeature([PasswordReset, User]),
     EmailModule,
   ],
   controllers: [AuthController],
-  providers: [AuthService],
-  exports: [JwtModule],
+  providers: [AuthService, AuthGuard, ConditionalAuthGuard],
+  exports: [JwtModule, AuthService, AuthGuard, ConditionalAuthGuard],
 })
 export class AuthModule {}