+++ src/notification/notification.controller.spec.ts
@@ -7,6 +7,7 @@ import { User } from '../dal/entity/user.entity';
 import { UserDevice } from '../dal/entity/userDevice.entity';
 import { NotificationController } from './notification.controller';
 import { NotificationService } from './notification.service';
+import { AuthService } from '../auth/auth.service';
 
 describe('NotificationController', () => {
   let controller: NotificationController;
@@ -39,6 +40,10 @@ describe('NotificationController', () => {
           },
         },
         JwtService,
+        {
+          provide: AuthService,
+          useValue: { verifyPwf: jest.fn() },
+        },
         {
           provide: getRepositoryToken(User),
           useValue: {